refactor(web): encapsulate main nav platform initialization
