+++ web/app/(commonLayout)/layout.tsx
@@ -1,9 +1,7 @@
 import * as React from 'react'
 import Zendesk from '@/app/components/base/zendesk'
 import MaintenanceNotice from '@/app/components/header/maintenance-notice'
-import MainNavLayout from '@/app/components/main-nav/layout'
-import { headers } from '@/next/headers'
-import { getPlatformFromUserAgent } from '@/utils/platform'
+import MainNavLayout from '@/app/components/main-nav/server'
 import { CommonLayoutGlobalMounts } from './global-mounts'
 import { ConsoleContextProviders, ConsoleRuntimeProviders } from './providers'
 
@@ -14,16 +12,13 @@ export default async function Layout({
   children: React.ReactNode
   detailSidebar: React.ReactNode
 }) {
-  const initialPlatform = getPlatformFromUserAgent((await headers()).get('user-agent'))
   return (
     <React.Fragment>
       <ConsoleRuntimeProviders>
         <div className="flex h-full flex-col overflow-hidden">
           <MaintenanceNotice />
           <ConsoleContextProviders>
-            <MainNavLayout detailSidebar={detailSidebar} initialPlatform={initialPlatform}>
-              {children}
-            </MainNavLayout>
+            <MainNavLayout detailSidebar={detailSidebar}>{children}</MainNavLayout>
             <CommonLayoutGlobalMounts />
           </ConsoleContextProviders>
         </div>
