+++ web/app/components/main-nav/__tests__/shortcut-platform.spec.ts
@@ -1,4 +1,4 @@
-import { getPlatformFromUserAgent } from './platform'
+import { getPlatformFromUserAgent } from '../shortcut-platform'
 
 describe('getPlatformFromUserAgent', () => {
   it.each([
