+++ web/app/components/main-nav/components/__tests__/search-button.spec.tsx
@@ -1,7 +1,7 @@
 import { act, waitFor, within } from '@testing-library/react'
 import { hydrateRoot } from 'react-dom/client'
 import { renderToString } from 'react-dom/server'
-import { getPlatformFromUserAgent } from '@/utils/platform'
+import { getPlatformFromUserAgent } from '../../shortcut-platform'
 import { MainNavSearchButton } from '../search-button'
 
 describe('MainNavSearchButton', () => {
