+++ web/app/components/main-nav/shortcut-platform.ts
